ses13

+++ Session13-BT MVC + SESSION/view/list_product.php
@@ -1,26 +1,33 @@
 <!DOCTYPE html>
 <html>
 <head>
-	<link rel="stylesheet" type="text/css" href="../css/style.css">
 	<title>List product</title>
 </head>
 <body>
 	<table>
-		<tr>
-			<th>No.</th>
-			<th>Name</th>
-			<th>Decription</th>
-			<th>Action</th>
-		</tr>
-		
 		<?php 
-			while($row = $list_product->fetch_assoc()) {
+		echo "<tr>";
+			echo "<th>No.</th>";
+			echo "<th>Name</th>";
+			echo "<th>Decription</th>";
+			echo "<th>Category</th>";
+			echo "<th>Image</th>";
+			if(isset($_SESSION['login']))
+			echo "<th>Action</th>";
+		echo "</tr>";
+		
+		
+			while($row = $list_category->fetch_assoc()) {
 				$id = $row['id'];
 				echo "<tr>";
 					echo "<td>".$id."</td>";
 					echo "<td>".$row['title']."</td>";
 					echo "<td>".$row['decription']."</td>";
+					echo "<td>".$row['TenDM']."</td>";
+					echo "<td><img src='./uploads/".$row['IMG']."'></td>";
+					if (isset($_SESSION['login'])) {
 					echo "<td><a href='index.php?action=delete_pro&id=".$id."'>DELETE</a>|<a href = 'index.php?action=edit_pro&id=".$id."'>EDIT</a></td>";
+					}
 				echo "</tr>";
 			}
 		?>
